moved potentially reusable pieces into lib
