+++ algorithms/policy_gradient/pg_config.py
@@ -5,10 +5,10 @@ from relaax.common.python.config.loaded_config import options
 class PGConfig(object):
 
     @classmethod
-    def preprocess(self):
+    def preprocess(cls):
         config = options.get('algorithm')
         # flatten state
-        config.state_size = np.prod(np.array(config.state_shape))
+        config.state_size = np.prod(np.array(config.state_size))
         return config
